+++ web/src/DeviceDetails.tsx
@@ -4,7 +4,7 @@ import { firestoreConnect } from "react-redux-firebase";
 import { compose } from "redux";
 import { Redirect } from "react-router-dom";
 import { DeviceState } from "./interfaces";
-import { deleteDevice } from "./store/deviceActions";
+import { deleteDevice } from "./store/systemActions";
 
 interface ParamType {
   id: Number;
